Moved files into modules directory/package
