+++ __init__.py
@@ -1,1 +1,3 @@
 from .modules.utils import *
+from .funcs import *
+from .api import *
