+++ cogs/general.py
@@ -209,19 +209,19 @@ class General(commands.Cog):
         await ctx.send()
        @commands.command(aliases=["Admin"])
     async def administrator(self,ctx):
-        await ctx.send("Note The Administrator's role within the SCP Foundation is often nebulous, but they always seem to have played a role in the organisation's founding, and are credited with writing the Mission Statement. Consistently associated with SCP-262. These things, at least, seem to be true, at least assuming the Administrator exists.\nIf reports are to be believed, each Administrator has had some anomalous traits which may be a factor of the position, including partial protection from reality shifts and extremely slow aging. Some reports even indicate that \"the Administrator\" is an anomalous entity contained by the Foundation.\nThe Administrator's position in relation to O5 Command is unclear. The Administrator may hold a position similar to an O5 Council member, and may be less powerful or more powerful depending on the situation and time. Because of all this, the Administrator often comes off seeming like a wild card, albeit a rather calm wild card.\nThere may have been multiple Administrators throughout the Foundation's history, but very few reports have indicated more than one at a time. Some reports claim that the Administrator has retired, others claim that the Administrator is deceased, and still others claim that \"the Administrator\" was only ever an alias used by another member of the O5 Council, possibly O5-1.\nThe Offical\nMale. European descent. American origin. Verified post-human lifespan; reported as appearing between mid-30s and mid-50s. No unusual appearance.\nNote: This version of the Administrator has been repeatedly confirmed as a false front, used as a cover for one or more of the Foundation's actual Administrators. May still be worth looking into further.\nDeliberately over-the-top reputation. Lack of aging made relatively public by way of universally repeated rumor. Associates with numerous major heads of government, representing Foundation interests to them. Said (rather insultingly) to have \"a penchant for suits, fine liquors and loose women\". An example from a typical report:\"He has been friendly with people of many political stances and is generally well liked if not respected. When he does become angered, even heads of state are quick to appease him and commonly offer sincere apologies.\"\nUsually said to have a number of over-the-top traits, such as never bringing his hands more than a few inches apart, and ordinarily being slow and sluggish and then displaying amazing feats of strength.\nF.Williams\nMale. European descent. American origin. Mid-30s. No unusual appearance.\nAccording to these reports, an individual known as \"F. Williams\" created the original [CORRUPTED]\n[FURTHER DATA LOST]\nWe're not even sure what that first initial is short for. We've found sources claiming it's \"Frederick\", \"Francis\", \"Fritz\", \"Friedrich\" or \"Franz\". Sometimes, the surname's something else or the name's \"William Fritz\" or \"Fritzwillis\", but the letters \"F. W.\" keep showing up.\nAgnes Peterson\nFemale. European descent. American origin. Verified post-human lifespan. Appears 50+. Old fashioned, wears horn-rim glasses, keeps hair in a bun. Very proper.\nOne of the original founders of the SCP Foundation. Pulled together the initial members into an organization; produced order from the initial chaos. Jokingly referred to as being very good at herding cats. Brisk in manner. Smart, ambitious, and good at knowing when to either create or ignore rules of order.\nKismet\nFemale. Middle Eastern descent. American origin. Verified post-human lifespan. Appears 50+. No unusual appearance.\nHer role is to track down and select each new and future member of O5 Command. She is a trained mathematician, and uses a form of statistics along with occultic procedures for her selection methods. She has personally selected every single member of O5 Command, in some cases years in advance. She occasionally also opts to choose their number, though which number is sometimes kept secret from all others except her and her staff until the new promotion is ready (presumably via a slot being opened).\nSpider\nMasculine. Descent/origin unknown. Head appears 30+. Unkempt white hair. Wears thick-lensed glasses and SCP-262 (Coat of Many Arms). Nonhuman body.\nApparently a human hybrid, or nonhuman entirely. Remarkable in appearance: A mostly human head, albeit with orange eyes. Body is completely nonhuman and hardly even bipedal, made of a twisted, weathered charcoal-like material, with atrophied natural limbs. A steel mesh covering holds the body inside the Coat of Many arms, and some of the many arms keep it in place. Other arms make up his \"arms\" and \"legs\", giving him either a generally humanoid look or a rather spidery look depending on how subtle he's attempting to be.\nApparently associated with the SCP Foundation since its inception, though always in a supporting role, not at the top of the power structure. Possibly owns the land Site-19 is built on. Only a few personnel outside of O5 Command have ever seen him. Softly spoken, but harshly punishes failure.\nReferences in reports claim this Administrator originates from \"The Plane Where Eyes Cannot Follow.\"\nThe Middleman\nMale. Unknown ethnicity. Unknown nationality. Unknown age.\nEfforts to obscure this man's identity go beyond what would be expected, even for someone of his position - I've seen video recordings of him and still couldn't describe what he actually looks like. It does seem that he was originally a file clerk for a government organisation dealing with the anomalous, but sold his secrets to the thirteen people that would become the O5 Council. Despite this, it's rumored that he's sided with the Foundation's Ethics Committee in opposing several O5 decisions. Reported dead multiple times, but never seems to stay that way, for some unknown reason.\nPossibly commands Mobile Task Force Rēsh-1 (\"Seat of Consciousness\"), although it's unclear if such a Task Force even exists.\nThe Archaeologist\nMale. European descent. English Origin. Appears 60+ (verified post-human lifespan).\nAccording to official records, Frederick William Abernathy engaged in some minor archaeological research, discovered nothing of note and died in obscurity. Unofficial sources indicate a much more eventful career, starting with the discovery of a Mesopotamian Goddess while excavating an Egyptian Tomb. Later encountered the entity known as the \"Guardian At the Gate\" and commanded to \"PREPARE\" (for what?). This was the impetus for the establishment of the Secret Society for Cultural Preservation, ~~the precursor~~ one possible precursor to the SCP Foundation. Current location and status unknown, but believed to still be in an influential position within the Foundation.\nCertain reports suggest that this individual (or someone similar) may have been succeeded by a female Administrator. If so, the original Administrator is likely deceased, although there are rumors that he's still able to advise the Foundation.\nThe Duke\nMale. European Descent. Austrian origin? Appears to be in his 40s, despite being over 300 years old. Human?\nKnown as Franz C. Williams. Inherited a title as an Austrian Duke within the Hapsburg Empire in the early 1700s. Emigrated to the USA during the Seven Years' War (also known as the First Occult War). Played an unknown role in the conflict, reason for leaving Europe unclear. Established a precursor to the Foundation from his mansion in rural Virginia. Arrived in the Americas with eleven others - the original twelve Overseers? Outlived multiple O5 Councils, continues to lead the Foundation in the present day.\nEthan Horowitz\nMale. Undetermined European ethnicity. Anomalous artifacts and techniques have extended his physical capabilities/longevity. Unknown true age.\nAn acquaintance of the Serpent prior to the Foundation's creation, and frequent face in the Wanderer's Library. Perpetually youthful, optimistic, charismatic, and charming. Creator of novel spellcraft. Formed the Foundation alongside O5-1, having been inspired by the Serpent's rejection from Heaven.\nHe drinks love so deeply, it's a wonder he hasn't drowned.\nToo many arms,\ntoo many faces,\nfor one man to carry.\nAssumed the form of Death in 2016.\nLed his ground-forces against humanity in 2███.\nAssented to the disembowelment of Heaven.\ni think\nhe simply\nlost sight\nof everything.")
+        await ctx.send()
       @commands.command()
     async def mtf(self,ctx):
-        await ctx.send("Mobile Task Forces (MTFs) are elite units comprised of personnel drawn from across the Foundation and are mobilized to deal with specific threats or situations that sometimes exceed the operational capacity or expertise of regular field personnel and — as their name suggests — may be relocated between facilities or locations as they are needed. Mobile Task Force personnel represent the \"best of the best\" of the Foundation.\nMobile Task Forces vary greatly in size, composition, and purpose. A battalion-strength combat-oriented task force trained to deal with highly aggressive anomalous entities may consist of hundreds of troops plus support personnel, vehicles, and equipment and can be deployed in whole or in part to deal with threats across the globe. However, a Mobile Task Force can also be a small, specialized intelligence-gathering or investigative task force that may have fewer than a dozen personnel if that is deemed sufficient to accomplish their goals.\nWhile in the field, task force members often pose as emergency responders, local or federal law enforcement, or military personnel appropriate to the region in which they are operating. Mobile Task Force Commanders can also request the assistance of local field units or personnel stationed at nearby Foundation facilities in order to accomplish their missions.\nOrganization\nEach unit is fundamentally structured in a way that best suits their intended purpose. While combat-oriented task forces may closely follow military hierarchy and organization, smaller units may have an informal or otherwise esoteric chain of command. As such, the responsibilities of the Mobile Task Force Commander (MTFC) for each particular task force can vary greatly; the commander for a large task force might focus on maintaining multiple teams and deploying them as necessary to each assigned operation, whereas the commander of a small team might deploy with their team and direct the operation from on location.\nSimilarly, the cohesion of each unit will vary as well. Some Mobile Task Forces consist of personnel who have trained and worked for many years or even decades together, whereas the personnel of a Mobile Task Force formed on a moment's notice to deal with a specific incident may know little more than each others' names and fields of expertise.\nCreation\nMobile Task Forces are typically commissioned as deemed necessary by the Foundation's Director of Task Forces, often with the direct approval of one or more O5 Council members. A significant number of Mobile Task Forces are created to deal with specific anomalies exhibiting traits that standard containment or response teams are unable to effectively counteract, though many were also created to pre-empt an emerging or theoretical threat.\nDeactivation\nMobile Task Forces created for the purpose of containing a particular anomaly are typically deactivated at the end of the recovery operation or when ongoing containment is deemed no longer necessary. Occasionally, such task forces remain operational if the expertise and experiences learned are considered useful for future incidents, but otherwise the task force will likely be disbanded and its personnel returned to their prior posts. Very rarely, a Mobile Task Force will also be disbanded if it suffers sufficient casualties to render it incapable of operation. In these cases, if the prior capability of that particular task force is deemed necessary, a new task force may be commissioned to replace it.")
+        await ctx.send()
       @commands.command(aliases=["RedRightHand,Alpha-1"])
     async def alpha1(self,ctx):
-        await ctx.send("Task Force Mission: Mobile Task Force Alpha-1 (Red Right Hand) is a task force that reports directly to the O5 Council and is used in situations that require the strictest operational security. The task force consists of the Foundation's best and most loyal operatives, and serves as the O5 Council's special operations unit. Further information regarding MTF Alpha-1 is classified Level 5.\nAssisting In Containment of Objects:\nSCP-001: Dead Men\nSCP-001: MAMJUL & KORAR\nSCP-2271\nSCP-3434\nSCP-3499\nSCP-3741\nSCP-3791\nSCP-4470\nSCP-5008\nSCP-5105\nSCP-5920\nSCP-7001\nSCP-7760")#add emblem
+        await ctx.send()#add emblem
       @commands.command(aliases=["PonyExpress,Alpha-4"])
     async def alpha4(self,ctx):
-        await ctx.send("Task Force Mission: Mobile Task Force Alpha-4 (Pony Express) consists primarily of personnel trained to act as undercover employees and specialize in tracking, intercepting, and securing anomalous objects sent through postal and package delivery services worldwide.\nAssisting In Containment of Objects:\nSCP-111\nSCP-130\nSCP-360\nSCP-2177\nSCP-2577\nSCP-2752\nSCP-3060\nSCP-3512\nSCP-3601\nSCP-4209\nSCP-5430") #add emblem
+        await ctx.send() #add emblem
       @commands.command(aliases=["Last Hope,Alpha-9"])
     async def alpha9(self,ctx):
-        await ctx.send("Task Force Mission: The reborn Omega-7.  Mobile Task Force Alpha-9 (Last Hope) is explicitly intended to train and utilize humanoid SCP objects in the field.\nUtilizing Objects:\nSCP-073\nSCP-105\nSCP-2913\nSCP-2987\nSCP-4051\nSCP-4494\nSCP-4818\nSCP-7414-A\nAssisting In Containment of Objects:\nSCP-5025") #add emblem
+        await ctx.send() #add emblem
       @commands.command(aliases=["Castaways,Beta-4"])
     async def beta4(self,ctx):
         await ctx.send("Task Force Mission: MTF Beta-4 (Castaways) is a task force created with the sole purpose of assisting and monitoring GoI-466 (Wilson's Wildlife Solutions) in their interactions with local fauna-based anomalies.\nAssisting In Containment of Objects:\nSCP-3465\nSCP-3466\nSCP-3467\nSCP-3577\nSCP-3676\nSCP-3879") #add emblem
