+++ cogs/generators/general.py
@@ -53,5 +53,10 @@ dossier0511=command("05-11: The Mailman\nMale. European descent. Eastern Europea
 dossier0512=command("05-12: The Accountant\n.\nMale. African descent. African origin. Appears to be mid-40s; almost certainly older. Large-bodied. Always dresses in extremely expensive tailored clothing with similarly expensive jewellery. Wears dark wrap-around glasses: smoked glass framed in ivory. May or may not be to hide his eyes. Likely occult significance.\nO5-12 plays the role of accountant for the Overseer Council. A mathematical genius; ensures all the Foundation's numbers add up. Has access to certain data that no other Foundation member does, for this purpose.\nPrecise, calm, practical. A strong believer in order. Every part of his daily route, and indeed every action he takes, right down to his drinking habit, is strictly scheduled.\n\"The Swallowed Tail\" source presents an alternate but still somewhat consistent picture of this O5-12. Not an original Council member, \"Omar Alawe\" was promoted following the death of the previous O5-12. His mathematical ability allows him to see correlations invisible to everyone else - not quite precognition, but may as well be. Some of the background is misinformation (likely including the photo — though I can't be sure, for obvious reasons), but trying to verify this information led down some interesting paths\n05-12: Adam [RETIRED]/[DECEASED]\nMale. European/Jewish descent. American origin. Somewhere between 130 and 200 years old. Appearance unknown.\nVery little information exists on O5-12's role within the Foundation, except that he was an exceptional researcher. Known primarily for fathering a stillborn child in the mid or late 1800s, and transforming her into SCP-321 in an ill-fated attempt to save her. Most information has been expunged from accessible Foundation records. Recovered information suggests a familial relation to Dr. Jack Bright. (Older brother? Father?)\n\"Adam\" is one of the few O5 Command members to retire. His retirement, by all accounts, was voluntary, which came as a surprise to me given his history. It appears that he was, in fact, the first ever O5 Command member to retire.\nInterestingly, it is unclear whether \"Adam\" was required to undergo the amnesic administration supposedly mandatory for retiring or removed O5 Command members. His location, unfortunately, is unknown to me.\nThere are also some rumors suggesting his connections to \"Department of Abnormalities\". Nature of this department and O5-12 connections to it requires more investigation.\nOther rumors, on the other hand, completely dismiss part of the above information, saying that he was one of the 12 original overseers, Convention on Preternatural Phenomena signatory — which created the modern Veil policy — and last general director of American Secure Containment Initiative before it merged into SCP Foundation in 1870. According to this version, \"Adam\" was also killed during the Great Factory Purge and his position was passed on to his descendant; probably \"The American\" who, according to this version, would have the position of O5-12 position instead of O5-6.\n05-12: The Contractor\nMale. African/European descent. American origin. Mid-40s. No unusual appearance.\nO5-12 was an emergency appointment following a previous assassination (or assassination attempt) during an O5 Council issue that required either an O5-12 or an O5-1, and O5-1 was unavailable.\nAfter this, he couldn't simply be demoted; therefore, he was assigned to side projects. He purchases land to build Foundation facilities on. He works with people who design Foundation facilities to make sure their needs are met; he works with local governments (often to request that military installations be built over actual Foundation sites as a cover), and he hires new Foundation personnel to implement such projects.\nAdditionally, O5-12 acts as an unofficial head of human services, requiring that Foundation personnel receive regular physical and mental checkups. Despite apparently caring a great deal about the average Foundation member, O5-12 is usually described as an obsessive, distrustful shut-in. His staff is very small, and they do very little.\n05-12: The Trainman/The Conductor [DEFECTED]\nMale. European descent. American origin. Appearance unknown.\nAssisted O5-11 in developing many of the Foundation's early containment procedures. Approved of weaponization of SCP objects. Supported extensive research into the anomalous.\nOne of O5-12's significant roles was overseeing the rotation of SCPs between Foundation facilities. This rotation was publicly aimed at getting new perspectives on SCP objects for research and effective containment purposes. Less public, but no less important, was the secondary purpose of preventing any group of researchers from working with humanoid SCPs for too long.\nInitially resisted defecting from the Foundation, but was strong-armed into joining the Insurgency by his wife, O5-11. His position of control over SCP rotation and transportation allowed the Insurgency to make off with a very high number of SCPs (particularly humanoid SCPs) during the schism. Additionally, his popularity with Foundation generals who desired weaponization of SCP objects gave the Insurgency a significant power boost during the Foundation Civil War.\n05-12: The Physician\nMale. East Asian descent. Chinese origin. Appears to be 70. Formerly affiliated with the Abnormality Institute (異學會; yì xué huì), sometimes called 'Chinese Abnormality Institute' (中華異學會; zhōng huá yì xué huì). Signatory of the Forbidden City Convention.\nBelieved to have been invited to join the O5 Council only due to his knowledge of a \"memory-altering drug\" (i.e. amnestics). Collaborated with the Chinese Nationalist Army to eliminate (betray?) the Meng clan; rationale is to monopolise amnestics for the Foundation.\nCoined the term 'Safe'. Rudimentary knowledge of English.\nNote: Amnestics might have undergone potentially radical advancements over the years. O5-12's statements regarding amnestics might not be applicable.\n05-12: Cyrus\nO5-12 is a brain in a jar.\nFrom the reports I've heard, his body was failing, and the other O5 council members wanted to keep him around as an advisor. They had the technology to do it, so they went ahead and preserved his mind.\nIt is unclear why the Foundation specifically chose this method, rather than one of the many other ways of prolonging life at their disposal.",aliases=["dossier05-12"])
 dossier0513=command("Of note A number of reports have indicated that the Foundation did not start with 13 O5 Council members. Most reports indicate that the original O5 Council numbered 12, and some indicate it still numbers only 12, with number 13 a work of disinformation.\nOther reports indicate an O5-13 was added after the fact due to the number 13 gaining modern occult significance and therefore power. And in practical terms, a tiebreaker vote was required once too often.\nSome sources indicate that there are more than thirteen members of the Overwatch Council, including mention of an O5-14 assigned to specific SCP-001 objects, or seizing power through anomalous means.\nSeveral of these options may be true. These reports reflect this lack of clarity.\n05-13: The Tiebreaker\nO5-13 does not exist. Everyone outside the O5 Council is told he exists, and memos are created under O5-13's name, but O5-13 exists only as a tiebreaker vote.\nO5-13's vote is rotated between each O5 Council member, starting at 1, and passed up a number each time a tiebreaker vote is required and used.\nHolding O5-13's vote is required to access a few selected databases. This knowledge may be attached to O5-13's vote (and wiped from the holder's memory when the vote is passed on), or it may simply comprise access to 13's unique databases.\n05-13: Tamlin\nEither male or gender-fluid (reports differ, may be relevant to anomalous traits). Unknown descent (reports differ, may be relevant to anomalous traits). Unknown origin - possibly Middle Eastern? Age unknown. Tall, red-headed.\nPer these reports, the \"Tiebreaker\" reports are accurate. However, O5-13 does exist, despite his vote not being held by himself most of the time. Original identity \"Dr. Joseph Tamlin\" or possibly \"Yoshua bin Yosef\" - unclear if these are alternative names for the same person, or different people holding the same position. \"Tamlin\" may (or may not) be more of a title than a specific individual - he/they continue to elude us.\nHolds only an advisory position by necessity. Is an anomalously altered human existing in a temporally destabilized state or location; partially \"unstuck in time\". Acts as a fail-safe for restoring 13's vote to its proper bearer, should it be compromised. Reserves the right to vote, but only exercises this right on pivotal decisions.\nSome reports (not all) indicate drastic shifts in appearance, race, and gender, but the presence of red hair (whether natural or colored) remains consistent. Unknown relevance; possible perception alteration issues or literal shape-shifting.\nOccupies a room (or rooms) in \"Tamlin House\", a hidden location related to SCP-1590, accessible only through a number of esoteric methods. Appears to be bound to the location. Does not usually participate directly in O5 Council meetings, but observes most of them (not necessarily synchronously).\n\"Shows what you know.\"\n05-13: The Meddler\nMale. European descent. American origin. Mid-30s. No unusual appearance.\nMost reports of O5-13 result from the conflicts he starts or becomes involved in, due to inter-office gossip from other O5's personal staff. No one outside the O5 Council appears to be aware of this O5 member's role in the Foundation; the role he generally seems to play in practice is interfering with other O5 members' projects. Known for ranting about how other O5 members (and lower-ranked Foundation personnel) are not running operations correctly.\nNot a particularly popular O5 Council member. Surprisingly, reports do not indicate many assassination attempts.\n05-13: Death\nGender unknown. Descent unknown. Origin unknown. Age unknown. Appearance unknown.\nOne of the more outlandish reports on a possible O5-13; claims that Death itself was bound by the O5 Council in the early 19th century. Per these reports, the slot of O5-13 was created for Death in the first place. Euphemistically referred to as \"The Other Overseer\".\nDeath is both a member and a prisoner of the O5 Council; it can cast a vote as normal. However, it is imprisoned by the rest of the Council, and by a majority vote of the O5 Council, the Council can control it. Death apparently otherwise carries out its functions as normal.\nSCP-006 does not exist; the longevity of O5s actually stems from O5-13. By majority vote, they can order O5-13 to stay its hand from any Council member — or indeed anyone in the world, though using this power for non-Council members seems to temporarily weaken Death's chains, and is therefore rarely used.\nWhen used, it is typically on a mass scale — SCP-2000 and similar anomalous items also do not exist, and are used as cover for every time that Council has voted to utilized O5-13's power to return most of the world's population back to life.\nThese reports may, or may not, conflict with other claims that the O5 Council's immortality comes from imprisoning Death, or from residing in a location shielded from the processes that cause mortality. Alternative reports claim that the Administrator, not O5-13, has assumed the role of Death.\nUnknown connection, if any, to SCP-2935.\n\"The Swallowed Tail\" source also identifies Death as having been bound to the Council by a bargain, in exchange for the life and Council seat of one Felix Carter, the former O5-13 (\"The Meddler\"). SCP-006 exists but is insufficient for the Council's purposes, and long since exhausted. The body of Felix Carter, which grants Death a council seat, is preserved somewhere in the South Atlantic.\nThis report contradicts the specifics of other reports claiming O5-13 to be \"Death\", but we can confirm that there is some form of anomaly in the South Atlantic, with antimemetic properties matching the description given in this report. Something is being hidden here.\n05-13: AI Complex\n[Personal traits not applicable]\nO5-13 is an artificial intelligence construct. This AI construct (using a variety of anomalous technological and non-technological methods) is supposedly able to take all factors relevant to the Foundation's goals and make a recommendation on any major course of action.\nDue to certain consequences (whether occult, resourced-based, or simply bureaucratic) of making use of this AI construct, it is only employed when the O5 Council must break a tie on a non-trivial matter.\nFurther details have been remarkably hard to attain even by esoteric methods. Whether or not this version of O5-13 exists, there are secrets behind this thing that the Foundation wishes to keep hidden.\n05-13: The Activist\nMale. South Asian descent. Sikh origin. Appears to be about 50. Formerly affiliated with 0th Anti-Cult Regiment, a colonial regiment in British India. Signatory of the Forbidden City Convention if \"The Persian\" is not present.\nMajor advocate for a meritocratic administration within the Foundation. Known to have formed an informal voting bloc with O5-11 and O5-12 to pursue racial equality within the Foundation.\nAlso known to have voted with O5-1 for many operational matters. Known O5-1 prior to the Forbidden City Convention. Perhaps a puppet for O5-1 to gain an additional vote in the O5 Council? Perhaps in exchange for said meritocratic administration?\n05-13: The Persian\nMale. Middle Eastern descent. Iranian origin. Appears to be about 70. Shia Muslim. Formerly an independent consultant on anomalous matters for the Ottoman Empire. Signatory of the Forbidden City Convention, if \"The Activist\" is not present.\nKnows O5-7 prior to the Forbidden City Convention. Holds strong suspicion towards O5-1.\nIn charge of the 'Autonomous Administration of the Greater Middle East', a sub-organisation of the Foundation based in the Middle East. Him and the AAGME do not exist in realities where \"The Activist\" exists. In fact, I can't seem to find the Office for the Reclamation of Islamic Artefacts or its equivalent in realities where \"The Persian\" takes up O5-13. AAGME and ORIA – counterparts???\nResembles ORIA founding member, Jibril Mani, in realities where \"The Activist\" is present instead.\n05-13: The Pedestrian\nMale. Latin American descent. Late-50s. No unusual appearance.\nO5-13's special connection to the anomalous gives him a perspective no other council member could begin to fathom. While Normalcy Confirmation meetings require eleven members for quorum, no meeting is allowed to proceed without O5-13's attendance.\nAdditionally, he is the only member of the original O5 council to survive the Caesar incident. Investigation into the nature of the incident and O5-13 is still ongoing.",aliases=["dossier05-13"])
 staffofthe05=command("O5 Command possesses their own Mobile Task Force, MTF Alpha-01, the Red Right Hand. They alternately act as the Foundation researcher's best friend and an outright medieval inquisition, complete with 'disappearing' people who the O5s disfavor.\nBesides MTF-Alpha-01, each O5 has their own personal staff and security. Usually O5 personal staff have Level 5 (\"Thaumiel\") access to Foundation archives (though they still need to be added to Special Access Programs, and do not have the \"blanket\" access that O5s do). Some of the O5's personal staff and secretaries are among the most powerful people in the Foundation.\nO5 Council members sometimes fixate on certain Foundation researchers. Many of the most well-known Foundation doctors and agents have been under the eye of one or more O5 members, willingly or not. This does not always end well for those favored in this way.\nFactotum\nAccording to some reports, the personal staff of the O5 Council are known as the Factotum. These Foundation operatives are selected for absolute loyalty, and act as personal assistants, bodyguards and decoys. Many encounters with an \"O5\" are actually mediated by a Factotum, dressed and disguised to perfectly resemble their assigned Council Member.\mReports ascribe some of these staff certain anomalous abilities, and the O5 Command reportedly approves this anomalous alteration to allow them to act as better decoys or bodyguards, directly circumventing official Foundation procedure. This makes it almost impossible to tell a decoy apart from the real thing.\nOf course, if they want to mislead, there's no reason to make this a perfect resemblance. For all we know, every one of the O5s described in this dossier are mere decoys, and the real O5 Council remains unknown.\nDelegation\nAlthough O5 Command holds tightly to its power over the Foundation, they do delegate many tasks. The exact command structure of the Foundation is unclear, but references have been made to the body known variously as the Site Directors’ Council, the Site Directors' Executive Committee of the Whole or the O4 Council (not to be confused with the O4 Command in charge of Lockdown Procedures). This organisation is assembled for matters requiring a broader perspective, or considered too minor for the O5 Council yet too significant or far reaching for any individual to decide. Under certain circumstances, this body may even have the power to depose and appoint O5 Council members, and it seems designed to ensure that the Foundation can continue functioning in the event that anything happens to the O5 Council.\nRegional branches of the SCP Foundation have their own equivalent bodies with full clearance for all anomalies under their regional jurisdiction, acting as an O5 equivalent for most intents and purposes, although still subordinate to the O5 Council. This includes the German Directors Council (known as the O4 Council), the Superintendence of the Italian Branch (The S5), and the Lusophone Board of Directors (The CL5).",aliases=["05staff"])
-infoaboutthe05=command("Appointing new O5s is rare to the point of being legendary, at least publicly. Rumors persist — potentially true — that some or all of the O5 Council members are inhuman or possess supernatural powers. Rumors also persist that they are unkillable, but this is at least partially untrue, given that several are known to be deceased.\nAs far as I can tell (and I may be wrong), all O5 Council members are human in origin, even if altered after the fact. A minority of the O5 Council may have anomalous abilities or traits, if reports are to be believed (in particular reports on O5-1 and O5-2), but none of the O5 Council are reality benders, and most of them are simply human… but humans with access to the Water of Life, and power beyond imagining. Far, far, far too much power.\nThe O5s rarely intervene in the day-to-day processes of the Foundation, or the containment of SCP objects. They tend to keep an eye on Keter-Class objects, and practically micromanage Thaumiels and Apollyons.\nBut sometimes — rarely — they'll jump into an apparently random Euclid- or Safe-Class object's documentation, and make some summary judgment with or without explanation. After all, no explanation is required of them.\nO5 Council members are officially barred from coming into contact with any anomalous entities. This is, naturally, not strictly true in practice. At minimum, all O5 Council members of sufficient tenure have access to SCP-006, a \"fountain of youth\", and likely to other means of prolonging their lifespan as well. And that aside, O5s tend to be able to bend rules, even their own rules.\nSome reports suggest that O5s regularly switch numbers with each other, either ad hoc or on a standardized basis. They make common use of body doubles and other security and disinformation techniques.\nThe Meeting Place of the O5 Council is frequently referred to as Protected Site-001, and one or more O5 Council Members may reside there most or all of the time. Alternative reports have described this headquarters as Building T-01 or O5 Command instead. Some reports claim that at least one O5 is in space at all times.\nOne particularly interesting report indicates that O5 Command members all wear an article of clothing and/or fetish object as a symbol of their office, with this totem being passed on to their successors. Possibly fashioned from the corpse of the most powerful reality bender to ever live, each object can completely stabilize reality around an O5 member. (Unclear how long this lasts, or if it is a permanent low-level effect.) Also said to grant additional abilities; precisely what these are remains unclear, as does the truth of this report.\nAnother set of reports claim that a number of humanoid SCPs are family members of O5 Council members. It is possible that fictional anomalous documentation was drawn up for them to justify keeping them under the best protection possible. There are also claims that some O5s have gone as far as to actually anomalously alter their family members to permanently justify Foundation protection via containment.\nThe O5 Council appears to choose their own members should a seat become vacant, although under certain circumstances they may become accountable to the Senior staff of the Foundation.\nThe exact relationship between O5 Command and the Foundation's Ethics Committee is unclear. I used to think that the Committee either didn't exist or just rubber-stamped the Council's decisions, but some reports suggest that the Ethics Committee has the power to overrule or even depose the Council. Who's really in charge here?\nThere are a few reports of an unofficial \"O5-0\", a former associate of the original O5 Council Members who left the Foundation due to ideological differences. Much like the official members of the Overwatch Council, I've encountered multiple contradictory descriptions of this \"O5-0\" - sorry, that seems to be a trend. They certainly sound like they'd be an interesting person to meet, if I ever get the opportunity…\nOne source bears notable mention: the reports, sources, and personas collectively referred to as \"The Swallowed Tail\". Please be warned before looking into this report: we acquired access to this documentation through the heart of the Chaos Insurgency, and they seem to have gotten most of their intel from a journal written by a Global Occult Coalition agent that later defected to the Foundation. If you're unfamiliar with the risks of Insurgency dealings, avoid looking at this, if you value your own life.\nWe initially thought this report to be Foundation disinformation (perhaps even targeted at me/us, the creator/creators of this dossier), but when we became aware of greater context, we realized we could draw zero conclusions.\nIs there a multiversal thread issue here that we're not aware of? A disconnection followed by an unknown bridge? An undescribed resonance? Something so obvious, too close to ourselves that we cannot see it? Something Outside, and dangerous?\nI/We really have to look into this.\nIf you don't know what you're doing, remember this: there are many worlds, and there are as many ends of the world. Some things you're better off not knowing. The number of these things may or may not be zero. Help, or don't.",aliases=["05 info"])
+infoaboutthe05=command("Appointing new O5s is rare to the point of being legendary, at least publicly. Rumors persist — potentially true — that some or all of the O5 Council members are inhuman or possess supernatural powers. Rumors also persist that they are unkillable, but this is at least partially untrue, given that several are known to be deceased.\nAs far as I can tell (and I may be wrong), all O5 Council members are human in origin, even if altered after the fact. A minority of the O5 Council may have anomalous abilities or traits, if reports are to be believed (in particular reports on O5-1 and O5-2), but none of the O5 Council are reality benders, and most of them are simply human… but humans with access to the Water of Life, and power beyond imagining. Far, far, far too much power.\nThe O5s rarely intervene in the day-to-day processes of the Foundation, or the containment of SCP objects. They tend to keep an eye on Keter-Class objects, and practically micromanage Thaumiels and Apollyons.\nBut sometimes — rarely — they'll jump into an apparently random Euclid- or Safe-Class object's documentation, and make some summary judgment with or without explanation. After all, no explanation is required of them.\nO5 Council members are officially barred from coming into contact with any anomalous entities. This is, naturally, not strictly true in practice. At minimum, all O5 Council members of sufficient tenure have access to SCP-006, a \"fountain of youth\", and likely to other means of prolonging their lifespan as well. And that aside, O5s tend to be able to bend rules, even their own rules.\nSome reports suggest that O5s regularly switch numbers with each other, either ad hoc or on a standardized basis. They make common use of body doubles and other security and disinformation techniques.\nThe Meeting Place of the O5 Council is frequently referred to as Protected Site-001, and one or more O5 Council Members may reside there most or all of the time. Alternative reports have described this headquarters as Building T-01 or O5 Command instead. Some reports claim that at least one O5 is in space at all times.\nOne particularly interesting report indicates that O5 Command members all wear an article of clothing and/or fetish object as a symbol of their office, with this totem being passed on to their successors. Possibly fashioned from the corpse of the most powerful reality bender to ever live, each object can completely stabilize reality around an O5 member. (Unclear how long this lasts, or if it is a permanent low-level effect.) Also said to grant additional abilities; precisely what these are remains unclear, as does the truth of this report.\nAnother set of reports claim that a number of humanoid SCPs are family members of O5 Council members. It is possible that fictional anomalous documentation was drawn up for them to justify keeping them under the best protection possible. There are also claims that some O5s have gone as far as to actually anomalously alter their family members to permanently justify Foundation protection via containment.\nThe O5 Council appears to choose their own members should a seat become vacant, although under certain circumstances they may become accountable to the Senior staff of the Foundation.\nThe exact relationship between O5 Command and the Foundation's Ethics Committee is unclear. I used to think that the Committee either didn't exist or just rubber-stamped the Council's decisions, but some reports suggest that the Ethics Committee has the power to overrule or even depose the Council. Who's really in charge here?\nThere are a few reports of an unofficial \"O5-0\", a former associate of the original O5 Council Members who left the Foundation due to ideological differences. Much like the official members of the Overwatch Council, I've encountered multiple contradictory descriptions of this \"O5-0\" - sorry, that seems to be a trend. They certainly sound like they'd be an interesting person to meet, if I ever get the opportunity…\nOne source bears notable mention: the reports, sources, and personas collectively referred to as \"The Swallowed Tail\". Please be warned before looking into this report: we acquired access to this documentation through the heart of the Chaos Insurgency, and they seem to have gotten most of their intel from a journal written by a Global Occult Coalition agent that later defected to the Foundation. If you're unfamiliar with the risks of Insurgency dealings, avoid looking at this, if you value your own life.\nWe initially thought this report to be Foundation disinformation (perhaps even targeted at me/us, the creator/creators of this dossier), but when we became aware of greater context, we realized we could draw zero conclusions.\nIs there a multiversal thread issue here that we're not aware of? A disconnection followed by an unknown bridge? An undescribed resonance? Something so obvious, too close to ourselves that we cannot see it? Something Outside, and dangerous?\nI/We really have to look into this.\nIf you don't know what you're doing, remember this: there are many worlds, and there are as many ends of the world. Some things you're better off not knowing. The number of these things may or may not be zero. Help, or don't.",aliases=["05info"])
 alt05theories=command("Some well-sourced reports suggest a specific membership or structure for the O5 Council that is not indicated in the previous reports in this dossier. These reports follow.\nSome well-sourced reports suggest that most information in this dossier is an elaborate facade and that the nature of the O5 Council is significantly different than what we think. For example, reports that O5 members are extraterrestrials, angels/demons from Abrahamic mythology, or simply do not exist at all. These reports follow, as well.\nUsurpation\nThe conflicting descriptions of the O5 Council strongly suggest that the Council's membership has changed before, perhaps many times. Claims that the entire O5 Council has been overthrown and replaced at various points are harder to verify since no two stories are the same. Some claim that the coup came from within the council, consolidating all power under a single seat, while others claim that the O5 Council was deemed corrupt or obsolete and so were deposed or replaced by their subordinates within the Foundation, or even by external enemies. If any of these reports have any basis in reality, it is likely that some (all?) of the information in this dossier is outdated.\nIndividuality\nAccording to some of our sources, the O5 Council was long ago overthrown. After the Chaos Insurgency defection catastrophe, the SCP Foundation was severely lacking in top-level administrators. Desperate, the O5 Council used anomalous genetic engineering and other tools to create personnel capable of filling key Directorships and executive roles. This proved to be a mistake.\nOnce in power, these entities conspired to take control of the Foundation. Using their combined influence, they were able to secretly remove the council and replace it with themselves. Mismanagement due to this instability may cause the Foundation to make mistakes.\nContainment\nIf the purpose of the Foundation is to contain the anomalous, the alleged abnormality of the O5 Council may seem like a betrayal of their original mission. But what if the Foundation exists to contain its Overseers? What if all the security and all the mystery exists not to protect the Council, but to hide them from the world?\nPerhaps it's been this way the entire time.\nPuppet Masters\nAlthough the O5s and the Administrator are generally regarded as being committed to the SCP Foundations stated goal of preserving \"normalcy\", multiple reports allege that the O5 Council is secretly responsible for creating anomalies.\nAdditionally, there are descriptions of the Administrator playing a role in creating anomalies, and founding, running and collaborating with multiple Groups of Interest, despite these organisations generally being seen as adversaries and competitors to the Foundation. It is unclear whether the rest of the O5 Council is aware of this.",aliases=["Alternate05theories"])
+administrator=command("Note The Administrator's role within the SCP Foundation is often nebulous, but they always seem to have played a role in the organisation's founding, and are credited with writing the Mission Statement. Consistently associated with SCP-262. These things, at least, seem to be true, at least assuming the Administrator exists.\nIf reports are to be believed, each Administrator has had some anomalous traits which may be a factor of the position, including partial protection from reality shifts and extremely slow aging. Some reports even indicate that \"the Administrator\" is an anomalous entity contained by the Foundation.\nThe Administrator's position in relation to O5 Command is unclear. The Administrator may hold a position similar to an O5 Council member, and may be less powerful or more powerful depending on the situation and time. Because of all this, the Administrator often comes off seeming like a wild card, albeit a rather calm wild card.\nThere may have been multiple Administrators throughout the Foundation's history, but very few reports have indicated more than one at a time. Some reports claim that the Administrator has retired, others claim that the Administrator is deceased, and still others claim that \"the Administrator\" was only ever an alias used by another member of the O5 Council, possibly O5-1.\nThe Offical\nMale. European descent. American origin. Verified post-human lifespan; reported as appearing between mid-30s and mid-50s. No unusual appearance.\nNote: This version of the Administrator has been repeatedly confirmed as a false front, used as a cover for one or more of the Foundation's actual Administrators. May still be worth looking into further.\nDeliberately over-the-top reputation. Lack of aging made relatively public by way of universally repeated rumor. Associates with numerous major heads of government, representing Foundation interests to them. Said (rather insultingly) to have \"a penchant for suits, fine liquors and loose women\". An example from a typical report:\"He has been friendly with people of many political stances and is generally well liked if not respected. When he does become angered, even heads of state are quick to appease him and commonly offer sincere apologies.\"\nUsually said to have a number of over-the-top traits, such as never bringing his hands more than a few inches apart, and ordinarily being slow and sluggish and then displaying amazing feats of strength.\nF.Williams\nMale. European descent. American origin. Mid-30s. No unusual appearance.\nAccording to these reports, an individual known as \"F. Williams\" created the original [CORRUPTED]\n[FURTHER DATA LOST]\nWe're not even sure what that first initial is short for. We've found sources claiming it's \"Frederick\", \"Francis\", \"Fritz\", \"Friedrich\" or \"Franz\". Sometimes, the surname's something else or the name's \"William Fritz\" or \"Fritzwillis\", but the letters \"F. W.\" keep showing up.\nAgnes Peterson\nFemale. European descent. American origin. Verified post-human lifespan. Appears 50+. Old fashioned, wears horn-rim glasses, keeps hair in a bun. Very proper.\nOne of the original founders of the SCP Foundation. Pulled together the initial members into an organization; produced order from the initial chaos. Jokingly referred to as being very good at herding cats. Brisk in manner. Smart, ambitious, and good at knowing when to either create or ignore rules of order.\nKismet\nFemale. Middle Eastern descent. American origin. Verified post-human lifespan. Appears 50+. No unusual appearance.\nHer role is to track down and select each new and future member of O5 Command. She is a trained mathematician, and uses a form of statistics along with occultic procedures for her selection methods. She has personally selected every single member of O5 Command, in some cases years in advance. She occasionally also opts to choose their number, though which number is sometimes kept secret from all others except her and her staff until the new promotion is ready (presumably via a slot being opened).\nSpider\nMasculine. Descent/origin unknown. Head appears 30+. Unkempt white hair. Wears thick-lensed glasses and SCP-262 (Coat of Many Arms). Nonhuman body.\nApparently a human hybrid, or nonhuman entirely. Remarkable in appearance: A mostly human head, albeit with orange eyes. Body is completely nonhuman and hardly even bipedal, made of a twisted, weathered charcoal-like material, with atrophied natural limbs. A steel mesh covering holds the body inside the Coat of Many arms, and some of the many arms keep it in place. Other arms make up his \"arms\" and \"legs\", giving him either a generally humanoid look or a rather spidery look depending on how subtle he's attempting to be.\nApparently associated with the SCP Foundation since its inception, though always in a supporting role, not at the top of the power structure. Possibly owns the land Site-19 is built on. Only a few personnel outside of O5 Command have ever seen him. Softly spoken, but harshly punishes failure.\nReferences in reports claim this Administrator originates from \"The Plane Where Eyes Cannot Follow.\"\nThe Middleman\nMale. Unknown ethnicity. Unknown nationality. Unknown age.\nEfforts to obscure this man's identity go beyond what would be expected, even for someone of his position - I've seen video recordings of him and still couldn't describe what he actually looks like. It does seem that he was originally a file clerk for a government organisation dealing with the anomalous, but sold his secrets to the thirteen people that would become the O5 Council. Despite this, it's rumored that he's sided with the Foundation's Ethics Committee in opposing several O5 decisions. Reported dead multiple times, but never seems to stay that way, for some unknown reason.\nPossibly commands Mobile Task Force Rēsh-1 (\"Seat of Consciousness\"), although it's unclear if such a Task Force even exists.\nThe Archaeologist\nMale. European descent. English Origin. Appears 60+ (verified post-human lifespan).\nAccording to official records, Frederick William Abernathy engaged in some minor archaeological research, discovered nothing of note and died in obscurity. Unofficial sources indicate a much more eventful career, starting with the discovery of a Mesopotamian Goddess while excavating an Egyptian Tomb. Later encountered the entity known as the \"Guardian At the Gate\" and commanded to \"PREPARE\" (for what?). This was the impetus for the establishment of the Secret Society for Cultural Preservation, ~~the precursor~~ one possible precursor to the SCP Foundation. Current location and status unknown, but believed to still be in an influential position within the Foundation.\nCertain reports suggest that this individual (or someone similar) may have been succeeded by a female Administrator. If so, the original Administrator is likely deceased, although there are rumors that he's still able to advise the Foundation.\nThe Duke\nMale. European Descent. Austrian origin? Appears to be in his 40s, despite being over 300 years old. Human?\nKnown as Franz C. Williams. Inherited a title as an Austrian Duke within the Hapsburg Empire in the early 1700s. Emigrated to the USA during the Seven Years' War (also known as the First Occult War). Played an unknown role in the conflict, reason for leaving Europe unclear. Established a precursor to the Foundation from his mansion in rural Virginia. Arrived in the Americas with eleven others - the original twelve Overseers? Outlived multiple O5 Councils, continues to lead the Foundation in the present day.\nEthan Horowitz\nMale. Undetermined European ethnicity. Anomalous artifacts and techniques have extended his physical capabilities/longevity. Unknown true age.\nAn acquaintance of the Serpent prior to the Foundation's creation, and frequent face in the Wanderer's Library. Perpetually youthful, optimistic, charismatic, and charming. Creator of novel spellcraft. Formed the Foundation alongside O5-1, having been inspired by the Serpent's rejection from Heaven.\nHe drinks love so deeply, it's a wonder he hasn't drowned.\nToo many arms,\ntoo many faces,\nfor one man to carry.\nAssumed the form of Death in 2016.\nLed his ground-forces against humanity in 2███.\nAssented to the disembowelment of Heaven.\ni think\nhe simply\nlost sight\nof everything.",aliases=["admin","admindossier"])
+mtf=command("Mobile Task Forces (MTFs) are elite units comprised of personnel drawn from across the Foundation and are mobilized to deal with specific threats or situations that sometimes exceed the operational capacity or expertise of regular field personnel and — as their name suggests — may be relocated between facilities or locations as they are needed. Mobile Task Force personnel represent the \"best of the best\" of the Foundation.\nMobile Task Forces vary greatly in size, composition, and purpose. A battalion-strength combat-oriented task force trained to deal with highly aggressive anomalous entities may consist of hundreds of troops plus support personnel, vehicles, and equipment and can be deployed in whole or in part to deal with threats across the globe. However, a Mobile Task Force can also be a small, specialized intelligence-gathering or investigative task force that may have fewer than a dozen personnel if that is deemed sufficient to accomplish their goals.\nWhile in the field, task force members often pose as emergency responders, local or federal law enforcement, or military personnel appropriate to the region in which they are operating. Mobile Task Force Commanders can also request the assistance of local field units or personnel stationed at nearby Foundation facilities in order to accomplish their missions.\nOrganization\nEach unit is fundamentally structured in a way that best suits their intended purpose. While combat-oriented task forces may closely follow military hierarchy and organization, smaller units may have an informal or otherwise esoteric chain of command. As such, the responsibilities of the Mobile Task Force Commander (MTFC) for each particular task force can vary greatly; the commander for a large task force might focus on maintaining multiple teams and deploying them as necessary to each assigned operation, whereas the commander of a small team might deploy with their team and direct the operation from on location.\nSimilarly, the cohesion of each unit will vary as well. Some Mobile Task Forces consist of personnel who have trained and worked for many years or even decades together, whereas the personnel of a Mobile Task Force formed on a moment's notice to deal with a specific incident may know little more than each others' names and fields of expertise.\nCreation\nMobile Task Forces are typically commissioned as deemed necessary by the Foundation's Director of Task Forces, often with the direct approval of one or more O5 Council members. A significant number of Mobile Task Forces are created to deal with specific anomalies exhibiting traits that standard containment or response teams are unable to effectively counteract, though many were also created to pre-empt an emerging or theoretical threat.\nDeactivation\nMobile Task Forces created for the purpose of containing a particular anomaly are typically deactivated at the end of the recovery operation or when ongoing containment is deemed no longer necessary. Occasionally, such task forces remain operational if the expertise and experiences learned are considered useful for future incidents, but otherwise the task force will likely be disbanded and its personnel returned to their prior posts. Very rarely, a Mobile Task Force will also be disbanded if it suffers sufficient casualties to render it incapable of operation. In these cases, if the prior capability of that particular task force is deemed necessary, a new task force may be commissioned to replace it.")
+alpha1=command("Task Force Mission: Mobile Task Force Alpha-1 (Red Right Hand) is a task force that reports directly to the O5 Council and is used in situations that require the strictest operational security. The task force consists of the Foundation's best and most loyal operatives, and serves as the O5 Council's special operations unit. Further information regarding MTF Alpha-1 is classified Level 5.\nAssisting In Containment of Objects:\nSCP-001: Dead Men\nSCP-001: MAMJUL & KORAR\nSCP-2271\nSCP-3434\nSCP-3499\nSCP-3741\nSCP-3791\nSCP-4470\nSCP-5008\nSCP-5105\nSCP-5920\nSCP-7001\nSCP-7760\nhttps://media.discordapp.net/attachments/812131957692694528/1099121878343360573/small.png?width=300&height=300",aliases=["redrighthand","alpha-1"])
+alpha4=command("Task Force Mission: Mobile Task Force Alpha-4 (Pony Express) consists primarily of personnel trained to act as undercover employees and specialize in tracking, intercepting, and securing anomalous objects sent through postal and package delivery services worldwide.\nAssisting In Containment of Objects:\nSCP-111\nSCP-130\nSCP-360\nSCP-2177\nSCP-2577\nSCP-2752\nSCP-3060\nSCP-3512\nSCP-3601\nSCP-4209\nSCP-5430\nhttps://scp-wiki.wdfiles.com/local--resized-images/task-forces/Alpha4.png/small.jpg",aliases=["ponyexpress","alpha-4"])
+alpha9=command("Task Force Mission: The reborn Omega-7.  Mobile Task Force Alpha-9 (Last Hope) is explicitly intended to train and utilize humanoid SCP objects in the field.\nUtilizing Objects:\nSCP-073\nSCP-105\nSCP-2913\nSCP-2987\nSCP-4051\nSCP-4494\nSCP-4818\nSCP-7414-A\nAssisting In Containment of Objects:\nSCP-5025\n")
