+++ modules/__init__.py
@@ -1,1 +1,3 @@
 from .utils import *
+from .funcs import *
+from .api import *
